+++ ShiningInfomation/ShiningInfomation/Content/Di.css
@@ -1,0 +1,9 @@
+﻿
+
+    .di{height: 180px;width:100%; background-color: #424242;}
+    .di div{width: 378px;float:left;margin: 0 auto; }
+    .di div ul,h4{width:150px;margin: 0 auto; list-style: none;}
+    .di div h4{color: #fff;font-size: 18px;padding:20px 0;}
+    .di div ul li{list-style: none;line-height: 30px;}
+    .di div ul li a{color:#a6a6a6;}
+    .di div ul li a:hover{color: #fff;}
