+++ ShiningInfomation/ShiningInfomation/Content/Site.css
@@ -22,3 +22,4 @@ select,
 textarea {
     max-width: 280px;
 }
+
