+++ ShiningInfomation/ShiningInfomation/Views/Admin/Create.cshtml
@@ -2,6 +2,7 @@
 
 @{
     ViewBag.Title = "Create";
+    Layout = "~/Views/Shared/_LayoutAdmin.cshtml";
 }
