+++ ShiningInfomation/ShiningInfomation/Views/Admin/Details.cshtml
@@ -2,6 +2,7 @@
 
 @{
     ViewBag.Title = "Details";
+    Layout = "~/Views/Shared/_LayoutAdmin.cshtml";
 }
