+++ ShiningInfomation/ShiningInfomation/Views/EditInfo/Delete.cshtml
@@ -2,15 +2,16 @@
 
 @{
     ViewBag.Title = "Delete";
-    Layout = "~/Views/Shared/_LayoutAdmin.cshtml";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
+<h2>Delete</h2>
 
-
+<h3>Are you sure you want to delete this?</h3>
 <div>
-    <h3>删除学生信息</h3>
-    <hr/>
-    <dl class="dl-horizontal" style="">
+    <h4>StudentInfo</h4>
+    <hr />
+    <dl class="dl-horizontal">
         <dt>
             @Html.DisplayNameFor(model => model.StudentName)
         </dt>
@@ -68,16 +69,13 @@
         </dd>
 
     </dl>
-    <hr />
 
     @using (Html.BeginForm()) {
         @Html.AntiForgeryToken()
 
-        <div class="form-actions no-color" style="margin-left:300px">
-            <input type="submit" value="删除" class="btn btn-default" style="margin-bottom:10px" />
+        <div class="form-actions no-color">
+            <input type="submit" value="Delete" class="btn btn-default" /> |
+            @Html.ActionLink("Back to List", "Index")
         </div>
-
-        @Html.ActionLink("返回", "Index")
-
     }
 </div>
