+++ ShiningInfomation/ShiningInfomation/Views/EditInfo/Details.cshtml
@@ -1,16 +1,16 @@
 ﻿@model ShiningInfomation.Entity.StudentInfo
 
 @{
-    ViewBag.Title = "Delete";
-    Layout = "~/Views/Shared/_LayoutAdmin.cshtml";
+    ViewBag.Title = "Details";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
-
+<h2>Details</h2>
 
 <div>
-    <h3>删除学生信息</h3>
-    <hr/>
-    <dl class="dl-horizontal" style="">
+    <h4>StudentInfo</h4>
+    <hr />
+    <dl class="dl-horizontal">
         <dt>
             @Html.DisplayNameFor(model => model.StudentName)
         </dt>
@@ -68,16 +68,8 @@
         </dd>
 
     </dl>
-    <hr />
-
-    @using (Html.BeginForm()) {
-        @Html.AntiForgeryToken()
-
-        <div class="form-actions no-color" style="margin-left:300px">
-            <input type="submit" value="删除" class="btn btn-default" style="margin-bottom:10px" />
-        </div>
-
-        @Html.ActionLink("返回", "Index")
-
-    }
 </div>
+<p>
+    @Html.ActionLink("Edit", "Edit", new { id = Model.StudentID }) |
+    @Html.ActionLink("Back to List", "Index")
+</p>
