+++ ShiningInfomation/ShiningInfomation/Views/EditInfo/Edit.cshtml
@@ -2,9 +2,10 @@
 
 @{
     ViewBag.Title = "Edit";
-    Layout = "~/Views/Shared/_LayoutAdmin.cshtml";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
+<h2>Edit</h2>
 
 
 @using (Html.BeginForm())
@@ -12,8 +13,7 @@
     @Html.AntiForgeryToken()
     
     <div class="form-horizontal">
-        <br/>
-        <h3>学生信息</h3>
+        <h4>StudentInfo</h4>
         <hr />
         @Html.ValidationSummary(true, "", new { @class = "text-danger" })
         @Html.HiddenFor(model => model.StudentID)
@@ -76,14 +76,14 @@
 
         <div class="form-group">
             <div class="col-md-offset-2 col-md-10">
-                <input type="submit" value="保存" class="btn btn-default" />
+                <input type="submit" value="Save" class="btn btn-default" />
             </div>
         </div>
     </div>
 }
 
 <div>
-    @Html.ActionLink("返回", "Index")
+    @Html.ActionLink("Back to List", "Index")
 </div>
 
 @section Scripts {
