+++ ShiningInfomation/ShiningInfomation/Views/Shared/_LayoutAdmin.cshtml
@@ -21,8 +21,7 @@
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
                     <li>@Html.ActionLink("主页", "Index", "Index")</li>
-                    <li>@Html.ActionLink("搜索信息", "Search", "Search")</li>
-                    <li>@Html.ActionLink("更多", "Index", "Index")</li>
+                    <li>@Html.ActionLink("管理员页", "Index", "Admin")</li>
                 </ul>
             </div>
         </div>
@@ -40,4 +39,3 @@
     @RenderSection("scripts", required: false)
 </body>
 </html>
-
